Error Handling ,fluentvalidation,axios,toastify

+++ Application/Activities/Delete.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Errors;
 using Domain;
 using MediatR;
 using Persistence;
@@ -25,13 +27,18 @@ namespace Application.Activities
             }
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
-                Activity activity = new Activity { Id = request.Id };
+                var activity = await _context.Activities.FindAsync(request.Id);
+                if (activity == null)
+                {
+                    throw new RestException(HttpStatusCode.NotFound, new { Id = request.Id, activity = "Not Found" });
+                }
+
+                //Activity activity = new Activity { Id = request.Id };
                 _context.Activities.Remove(activity);
                 bool success = await _context.SaveChangesAsync() > 0;
-
                 if (!success)
                 {
-                    throw new Exception("Activity could not be Deleted");
+                    throw new RestException(HttpStatusCode.NotFound, new { Id = request.Id, activity = "Activity could not be deleted" });
                 }
 
                 return Unit.Value;
